Bit more refactor
