+++ src/scripts/vscode.ts
@@ -1,5 +1,13 @@
 import { provideVSCodeDesignSystem, vsCodeButton, vsCodeCheckbox, vsCodeDivider, vsCodeDropdown, vsCodeOption, vsCodeTextField } from '@vscode/webview-ui-toolkit';
+import { ClientMessage } from '../types/shared/messages';
 
-provideVSCodeDesignSystem().register(vsCodeButton(), vsCodeCheckbox(), vsCodeTextField(), vsCodeDivider(), vsCodeDropdown(), vsCodeOption());
+
+export function initialiseVSCode() {
+	provideVSCodeDesignSystem().register(vsCodeButton(), vsCodeCheckbox(), vsCodeTextField(), vsCodeDivider(), vsCodeDropdown(), vsCodeOption());
+}
 
 export const vscode = acquireVsCodeApi();
+
+export function sendMessage(message: ClientMessage) {
+	vscode.postMessage(message);
+}
